+++ @xen-orchestra/web/src/modules/vm/components/VmGuestToolsStatus.vue
